Factor snapfile de/serialization into a trait
