+++ src/snapshot.rs
@@ -1,12 +1,13 @@
 use std::error::Error;
 use std::fs;
-use std::io::{BufRead, BufReader, Write};
+use std::io::{BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use lazy_static::lazy_static;
 use serde::{Deserialize, Serialize};
 
+use super::loader::{DefaultSnapfileFormatter, SnapfileFormatter};
 use super::runtime::get_inline_snapshot_value;
 
 lazy_static! {
@@ -115,7 +116,7 @@ pub struct Snapshot {
     module_name: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     snapshot_name: Option<String>,
-    metadata: MetaData,
+    pub(crate) metadata: MetaData,
     snapshot: SnapshotContents,
 }
 
@@ -123,87 +124,7 @@ impl Snapshot {
     /// Loads a snapshot from a file.
     pub fn from_file<P: AsRef<Path>>(p: P) -> Result<Snapshot, Box<dyn Error>> {
         let mut f = BufReader::new(fs::File::open(p.as_ref())?);
-        let mut buf = String::new();
-
-        f.read_line(&mut buf)?;
-
-        // yaml format
-        let metadata = if buf.trim_end() == "---" {
-            loop {
-                let read = f.read_line(&mut buf)?;
-                if read == 0 {
-                    break;
-                }
-                if buf[buf.len() - read..].trim_end() == "---" {
-                    buf.truncate(buf.len() - read);
-                    break;
-                }
-            }
-            serde_yaml::from_str(&buf)?
-        // legacy format
-        } else {
-            let mut rv = MetaData::default();
-            loop {
-                buf.clear();
-                let read = f.read_line(&mut buf)?;
-                if read == 0 || buf.trim_end().is_empty() {
-                    buf.truncate(buf.len() - read);
-                    break;
-                }
-                let mut iter = buf.splitn(2, ':');
-                if let Some(key) = iter.next() {
-                    if let Some(value) = iter.next() {
-                        let value = value.trim();
-                        match key.to_lowercase().as_str() {
-                            "expression" => rv.expression = Some(value.to_string()),
-                            "source" => rv.source = Some(value.into()),
-                            _ => {}
-                        }
-                    }
-                }
-            }
-            rv
-        };
-
-        buf.clear();
-        for (idx, line) in f.lines().enumerate() {
-            let line = line?;
-            if idx > 0 {
-                buf.push('\n');
-            }
-            buf.push_str(&line);
-        }
-
-        let module_name = p
-            .as_ref()
-            .file_name()
-            .unwrap()
-            .to_str()
-            .unwrap_or("")
-            .split("__")
-            .next()
-            .unwrap_or("<unknown>")
-            .to_string();
-
-        let snapshot_name = p
-            .as_ref()
-            .file_name()
-            .unwrap()
-            .to_str()
-            .unwrap_or("")
-            .split('.')
-            .next()
-            .unwrap_or("")
-            .splitn(2, "__")
-            .nth(1)
-            .map(|x| x.to_string());
-
-        Ok(Snapshot::from_components(
-            module_name,
-            snapshot_name,
-            metadata,
-            buf.into(),
-        ))
+        DefaultSnapfileFormatter::deserialize(&mut f, p.as_ref().file_name().unwrap())
     }
 
     /// Creates an empty snapshot.
@@ -252,11 +173,7 @@ impl Snapshot {
             fs::create_dir_all(&folder)?;
         }
         let mut f = fs::File::create(&path)?;
-        serde_yaml::to_writer(&mut f, &self.metadata)?;
-        f.write_all(b"\n---\n")?;
-        f.write_all(self.contents_str().as_bytes())?;
-        f.write_all(b"\n")?;
-        Ok(())
+        DefaultSnapfileFormatter::serialize(self, &mut f)
     }
 }
